I believe this will fix the issue I have of it being " out of date"
Reminder: 2e'sCustomService is a pirate, but only on the tweak side. We don't support any kind of piracy, yet I accept him, and respect him. If you have a problem with us not hard-fighting this pirate, please DM John NemECis. 

the reasons to accept him, are good reasons. The reason he made his stance, is because how developers treated him. I personally stand behind him because his views regularly side the ones of Saurik imo. Good to mention too: We have a different view on how to respond on developers being jerks. Our method is to hope they change, his is to take their packages from them and hand them out for free.

again; We do not condone piracy. But we do respect his contributions, how controversial they are.


Notes about the update. We added a border, we added a few other things, and we made the ImagePicker (still broken, but looks cool when it responds.